@@ -1,6 +1,6 @@
 {
   "name": "desktop-idle",
-  "version": "1.3.0",
+  "version": "2.0.0",
   "main": "index.js",
   "license": "MIT",
   "description": "Node/Electron module to detect idle desktop users(users away from keyboard) in macOS, Windows, Linux, FreeBSD and OpenBSD",
@@ -35,12 +35,15 @@
     "openbsd"
   ],
   "engines": {
-    "node": ">=7.9.0"
+    "node": ">=22.14.0"
   },
   "devDependencies": {
-    "eslint": "^4.6.1",
-    "npm-watch": "^0.2.0",
-    "vows": "^0.8.1"
+    "eslint": "9.23.0",
+    "npm-watch": "0.13.0",
+    "vows": "0.8.3",
+    "npm-check": "6.0.1",
+    "node-gyp": "11.1.0",
+    "typescript": "5.8.2"
   },
   "watch": {
     "test_watch": {
@@ -55,6 +58,10 @@
   "scripts": {
     "test_watch": "npm install && npm test",
     "test": "node test/main_test.js",
-    "watch": "npm-watch"
+    "watch": "npm-watch",
+    "gyp-configure": "node-gyp configure",
+    "gyp-build": "node-gyp build",
+    "check-packages": "npx npm-check -u -E",
+    "refresh-packages": "rm -rf node_modules && rm -rf package-lock.json && npm i"
   }
 }
